Add files via upload
A simple Java class based on a game of guessing the number. The class is "GuessTheNumberGame" .
I imported classes that gives random numbers so the player loops through numbers till they guess. 
The variables lowerBound and upperBound allows the player to choose numbers higher and lower.
The variable secretNumber contains the number that should be guessed.
The boolean hasGuessedCorrectly becomes true when the player guessed right.
In the code is a constructor, a method "generateRandomNumber", a "play" method and the main method.